Revert "Add image chooser functionality (#4)"
This reverts commit a1a1836e0c85130594e9a1c6b4ddb8fde8faf549.
