+++ src/Program.cs
@@ -1,13 +1,10 @@
 ﻿using Avalonia;
-using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Logging.Serilog;
-using Avalonia.ReactiveUI;
 using Filters.Views;
-using System;
 
 namespace Filters
 {
-    public class Program
+    class Program
     {
         // Initialization code. Don't use any Avalonia, third-party APIs or any
         // SynchronizationContext-reliant code before AppMain is called: things aren't initialized
@@ -19,7 +16,6 @@ namespace Filters
         public static AppBuilder BuildAvaloniaApp()
             => AppBuilder.Configure<App>()
                 .UsePlatformDetect()
-                .LogToDebug()
-                .UseReactiveUI();
+                .LogToDebug();
     }
 }
