+++ src/filters.csproj
@@ -14,8 +14,6 @@
   <ItemGroup>
     <PackageReference Include="Avalonia" Version="0.9.9" />
     <PackageReference Include="Avalonia.Desktop" Version="0.9.9" />
-    <PackageReference Include="Avalonia.ReactiveUI" Version="0.9.9" />
     <PackageReference Include="Citrus.Avalonia" Version="1.2.6" />
-    <PackageReference Include="ReactiveUI" Version="11.4.17" />
   </ItemGroup>
 </Project>
