+++ src/views/App.xaml.cs
@@ -1,8 +1,6 @@
 using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
-using Filters.Models;
-using Filters.Views.Main;
 
 namespace Filters.Views
 {
@@ -19,10 +17,7 @@ namespace Filters.Views
         {
             if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
             {
-                desktop.MainWindow = new MainWindow()
-                {
-                    DataContext = new MainWindowModel()
-                };
+                desktop.MainWindow = new MainWindow();
             }
 
             base.OnFrameworkInitializationCompleted();
